+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 [Squint](https://github.com/squint-cljs/squint): Light-weight ClojureScript dialect
 
-## Unreleased
+## v0.8.138 (2025-03-04)
 
 - Fix [#603](https://github.com/squint-cljs/squint/issues/603): don't emit multiple returns
 - Drop constructor requirement for `defclass`